pyproject.toml
[tool.poetry]
name = "time-to-do"
version = "0.1.0"
description = "Task Tracker с SCRUM/Kanban методологиями для эффективной командной работы"
authors = ["Tomas <tomas@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
# Базовый Python - соответствует последней стабильной версии
python = "^3.13"

# Core FastAPI стек - актуальные версии 2026
fastapi = {extras = ["standard"], version = "^0.128.0"}  # ✅ Актуальная версия
uvicorn = {extras = ["standard"], version = "^0.40.0"}  # ✅ Актуальная версия
starlette = "^0.50.0"  # ✅ Совместимо с FastAPI 0.128.0

# База данных и ORM - актуальные версии
sqlalchemy = "^2.0.46"  # ✅ Актуальная версия
asyncpg = "^0.31.0"  # ✅ Актуальная версия для async PostgreSQL
alembic = "^1.18.3"  # ✅ Актуальная версия

# Валидация данных и настройки
pydantic = {extras = ["email"], version = "^2.12.5"}  # ✅ Актуальная версия
pydantic-settings = "^2.12.0"  # ✅ Обновлено до актуальной версии
email-validator = "^2.3.0"  # ✅ Добавлено для валидации email

# Аутентификация и безопасность - современные версии
pyjwt = {extras = ["crypto"], version = "^2.11.0"}  # ✅ Современная замена python-jose
bcrypt = "^4.3.0"  # ✅ Совместимо с проектом

# HTTP и multipart - совместимые версии
httpx = "^0.27.2"  # ✅ Обновлено для совместимости с fastapi[standard]
python-multipart = "^0.0.22"  # ✅ Актуальная версия

# Вспомогательные библиотеки - совместимые версии
redis = "^5.3.1"  # ✅ Совместимо с текущими зависимостями
aiofiles = "^23.2.1"  # ✅ Совместимо с проектом
jinja2 = "^3.1.6"  # ✅ Обновлено до последней версии
python-dotenv = "^1.2.1"  # ✅ Обновлено до последней версии

[tool.poetry.group.dev.dependencies]
# Тестирование - актуальные версии 2026
pytest = "^9.0.2"  # ✅ Актуальная версия
pytest-asyncio = "^1.3.0"  # ✅ Обновлено до последней версии
pytest-cov = "^7.0.0"  # ✅ Обновлено до последней версии
factory-boy = "^3.3.1"  # ✅ Factory Pattern для тестовых данных
hypothesis = "^6.108.0"  # ✅ Property-based testing

# Качество кода - лучшие инструменты 2026
black = "^26.1.0"  # ✅ Актуальная версия
ruff = "^0.14.14"  # ✅ Актуальная версия
mypy = "^1.19.1"  # ✅ Актуальная версия

# Инструменты разработки - обновленные версии
pre-commit = "^4.5.1"  # ✅ Обновлено до последней версии
platformdirs = "^4.5.1"  # ✅ Добавлено для совместимости
bandit = "^1.9.3"  # ✅ Инструмент проверки безопасности

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py312']  # Соответствует Python 3.12 для совместимости с Ruff
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.ruff]
target-version = "py312"  # Согласовано с остальной конфигурацией
line-length = 88

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.13"
check_untyped_defs = true
disallow_any_generics = false
disallow_incomplete_defs = true
disallow_untyped_defs = false
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = false

[[tool.mypy.overrides]]
module = [
    "sqlalchemy.*",
    "alembic.*",
]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = [
    "app.models.*",
]
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-ra -q --strict-markers --strict-config"
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]

[tool.coverage.run]
source = ["app"]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "*/__pycache__/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
]
